Sync Local Project
push for local project files.

+++ include/random.h
@@ -17,7 +17,7 @@ u16 Random2(void);
 #define ISO_RANDOMIZE2(val) (1103515245 * (val) + 12345)
 
 //Sets the initial seed value of the pseudorandom number generator
-void SeedRng(u16 seed);
+void SeedRng(u32 seed);
 void SeedRng2(u16 seed);
 
 #endif // GUARD_RANDOM_H
